+++ pydanticInput/__init__.py
@@ -1,1 +1,2 @@
+from pydanticInput.dispatch import type_dispatch  # noqa: F401
 from pydanticInput.main import Input  # noqa: F401
